Prepare RecombeeClient for Swift 6.0 concurency (#1)
* feat: prepare RecombeeClient for Swift 6.0 concurency

Mark necessary protocols and types as sendable. Use unchecked sendable for RecombeeClient. Rework AnyBinding to not use Any as Any cant be sendable by design, instead use more type safe way.

RecombeeClient `class` as a whole could be replaced with `actor` but then the idea of fire and forget `sendDetached` would miss a point as we would still need to await the function call in the client apps. This way I've only marked the class as `@unchecked Sendable` as the class itself does not have any mutating properties and there is no data race issues.

* fix: remove unnecessary (duplicated) protocol conformance

* swiftformat

---------

Co-authored-by: Ondrej Fiedler <ondrej.fiedler@recombee.com>